requirements.txt
#python 3.9.10
pycryptodomex 3.14.1